// Szakdolgozat_Frontend_Angular/szakdolgozat-frontend/src/app/models/communities_model.ts
export interface CommunitiesData {
    timestamp: Date;
    number_of_panels: number;
    panel_area_m2: number;
    category: string;
    consumption: number;
    air_temp: number;
    clearsky_dhi: number;
    clearsky_dni: number;
    clearsky_ghi: number;
    clearsky_gti: number;
    cloud_opacity: number;
    dhi: number;
    dni: number;
    ghi: number;
    gti: number;
    snow_soiling_rooftop: number;
    snow_soiling_ground: number;
    season: string;
  }